@@ -6,13 +6,12 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface CommentRepository extends JpaRepository<Comment, Long> {
-
-    Long countByPostId(Long postId);
-
     List<Comment> findByPost(Post post);
 }
