@@ -1,13 +1,12 @@
 package backend.mbti.dto.post;
 
 import lombok.*;
-
-import java.util.Date;
+import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 
 @Getter
 @Setter
 @RequiredArgsConstructor
-public class PostCreateRequest {
+public class  CreatePostRequest{
     private String title;
     private String optionA;
     private String optionB;
