@@ -1,20 +1,19 @@
 package backend.mbti.post;
 
 
-import backend.mbti.dto.post.PostCreateRequest;
+import backend.mbti.dto.post.CreatePostRequest;
 import backend.mbti.dto.post.PostResponse;
-import backend.mbti.dto.post.PostUpdateRequest;
-import backend.mbti.domain.post.Post;
+import backend.mbti.dto.post.UpdatePostRequest;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public interface PostService {
-    Post createPost(PostCreateRequest request, String username);
+    void createPost(CreatePostRequest createPostRequest, String username);
     List<PostResponse> viewPostList();
     PostResponse viewPost(Long postId);
-    Post updatePost(Long postId, PostUpdateRequest request, String username);
+    void updatePost(Long postId, UpdatePostRequest request, String username);
     void deletePost(Long postId, String username);
     void likePost(Long postId, String username);
 }
